# docs/api.yaml
openapi: 3.1.0
info:
  title: ReEngageAI Core API
  version: 1.0.0
  description: Commerce, behaviour ingestion, recommendation, notification, and analytics API.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /actuator/health:
    get:
      security: []
      summary: Aggregated service health
      responses:
        "200": {description: Healthy}
  /api/v1/auth/register:
    post:
      security: []
      summary: Register a customer
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RegisterRequest"}
      responses:
        "200":
          description: JWT and identity
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
        "400": {$ref: "#/components/responses/Problem"}
  /api/v1/auth/login:
    post:
      security: []
      summary: Authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: JWT and identity
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
        "400": {$ref: "#/components/responses/Problem"}
  /api/v1/auth/me:
    get:
      summary: Current identity
      responses:
        "200": {description: Identity}
        "401": {$ref: "#/components/responses/Problem"}
  /api/v1/products:
    get:
      security: []
      summary: Search and filter products
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: brand, in: query, schema: {type: string}}
        - {name: minPrice, in: query, schema: {type: integer, minimum: 0}}
        - {name: maxPrice, in: query, schema: {type: integer, minimum: 0}}
      responses:
        "200":
          description: Products and facets
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProductCollection"}
  /api/v1/products/{productId}:
    get:
      security: []
      summary: Product detail
      parameters:
        - {name: productId, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Product"}
        "404": {$ref: "#/components/responses/Problem"}
  /api/v1/events/batch:
    post:
      security: []
      summary: Idempotently ingest behaviour events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: {$ref: "#/components/schemas/BehaviourEvent"}
      responses:
        "202":
          description: Ingestion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: {type: integer}
                  duplicates: {type: integer}
                  rejected: {type: array, items: {type: object}}
        "400": {$ref: "#/components/responses/Problem"}
  /api/v1/cart:
    get:
      summary: Current user's cart
      responses:
        "200": {$ref: "#/components/responses/Cart"}
  /api/v1/cart/items/{productId}:
    put:
      summary: Add or update a cart item
      parameters:
        - {name: productId, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: {type: integer, minimum: 1}
      responses:
        "200": {$ref: "#/components/responses/Cart"}
    delete:
      summary: Remove a cart item
      parameters:
        - {name: productId, in: path, required: true, schema: {type: string}}
      responses:
        "200": {$ref: "#/components/responses/Cart"}
  /api/v1/checkout:
    post:
      summary: Atomically create an order and cancel pending reminders
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, maxLength: 120}
      responses:
        "200": {description: New or replayed order}
        "400": {$ref: "#/components/responses/Problem"}
  /api/v1/orders:
    get:
      summary: Current user's order history
      responses:
        "200": {description: Orders}
  /api/v1/profile:
    get:
      summary: Current user's computed profile and intent
      responses:
        "200": {description: Profile}
  /api/v1/recommendations:
    get:
      summary: Ranked business-filtered recommendations
      parameters:
        - {name: anchor, in: query, schema: {type: string}}
      responses:
        "200": {description: Recommendations}
  /api/v1/preferences:
    get:
      summary: Consent, channel, timezone, and quiet hours
      responses:
        "200": {description: Preferences}
    put:
      summary: Update preferences and cancel jobs on full opt-out
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PreferenceRequest"}
      responses:
        "200": {description: Updated preferences}
  /api/v1/notifications:
    get:
      summary: Current user's notification history
      responses:
        "200": {description: Notifications}
  /api/v1/notifications/{notificationId}/open:
    post:
      summary: Record an open
      parameters:
        - {name: notificationId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Recorded}
  /api/v1/notifications/{notificationId}/click:
    post:
      summary: Record a click
      parameters:
        - {name: notificationId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Recorded}
  /api/v1/admin/analytics:
    get:
      summary: Operator metrics, activity, queue, and variants
      responses:
        "200": {description: Analytics snapshot}
        "403": {$ref: "#/components/responses/Problem"}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: RFC-style application problem
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Problem"}
    Cart:
      description: Cart contents and total
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Cart"}
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 8, maxLength: 72}
        displayName: {type: string, maxLength: 80}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}
    AuthResponse:
      type: object
      required: [accessToken, userId, email, role]
      properties:
        accessToken: {type: string}
        userId: {type: string, format: uuid}
        email: {type: string, format: email}
        role: {type: string, enum: [CUSTOMER, ADMIN]}
    BehaviourEvent:
      type: object
      required: [eventId, sessionId, eventType, timestamp]
      anyOf:
        - required: [userId]
        - required: [anonymousId]
      properties:
        eventId: {type: string, format: uuid}
        userId: {type: string, format: uuid}
        anonymousId: {type: string, maxLength: 160}
        sessionId: {type: string, format: uuid}
        eventType:
          type: string
          enum: [PAGE_VIEWED, PRODUCT_VIEWED, SEARCH_PERFORMED, FILTER_APPLIED, PRODUCT_COMPARED, ADD_TO_CART, REMOVE_FROM_CART, CHECKOUT_STARTED, PURCHASE_COMPLETED, TIME_SPENT, SESSION_STARTED, SESSION_ENDED, NOTIFICATION_CLICKED]
        productId: {type: string}
        timestamp: {type: string, format: date-time}
        sourcePage: {type: string}
        device: {type: object, additionalProperties: true}
        metadata: {type: object, additionalProperties: true}
        schemaVersion: {type: integer, default: 1}
    Product:
      type: object
      required: [id, name, brand, category, priceInr, stock]
      properties:
        id: {type: string}
        name: {type: string}
        brand: {type: string}
        category: {type: string}
        description: {type: string}
        priceInr: {type: integer}
        rating: {type: number}
        stock: {type: integer}
        qualityScore: {type: number}
        features: {type: string}
        accent: {type: string}
        imageKey: {type: string}
    ProductCollection:
      type: object
      properties:
        products: {type: array, items: {$ref: "#/components/schemas/Product"}}
        categories: {type: array, items: {type: string}}
        brands: {type: array, items: {type: string}}
    Cart:
      type: object
      properties:
        items: {type: array, items: {type: object}}
        totalInr: {type: integer}
    PreferenceRequest:
      type: object
      required: [emailOptIn, whatsappOptIn, preferredChannel, timezone, quietStart, quietEnd]
      properties:
        emailOptIn: {type: boolean}
        whatsappOptIn: {type: boolean}
        preferredChannel: {type: string, enum: [EMAIL, WHATSAPP]}
        timezone: {type: string, example: Asia/Kolkata}
        quietStart: {type: integer, minimum: 0, maximum: 23}
        quietEnd: {type: integer, minimum: 0, maximum: 23}
        phoneE164:
          type: string
          pattern: '^\+[1-9]\d{7,14}$'
          example: "+919876543210"
    Problem:
      type: object
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        error: {type: string}
        message: {type: string}
        path: {type: string}